refactor(static-analysis): keep only required memory checks
Amp-Thread-ID: https://ampcode.com/threads/T-019fc95c-d9c9-716e-8ca4-66d5fbd7812c
